21-6-2020

+++ app/Http/Controllers/Api/V1/Sale/SaleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api\V1\Sale;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Sale\StoreSale;
+use App\Http\Resources\product\WebProductResource;
 use App\Http\Resources\Sale\SaleResource;
 use App\Products;
 use App\Sales;
@@ -14,11 +15,20 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class SaleController extends Controller
 {
 
     use ApiResponser;
+
+    public function __construct()
+    {
+
+        $this->middleware('checkIfUserHasProduct')->only('show');
+        $this->middleware(['permission:add_product','checkIfUserHasProduct'])->only('store');
+        $this->middleware(['permission:delete_product','checkIfUserHasProduct'])->only('destroy');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -27,42 +37,72 @@ class SaleController extends Controller
     public function index()
     {
         if (request()->expectsJson() && request()->acceptsJson()){
-            $sales = Sales::all();
-            return $this->showCollection(SaleResource::collection($sales));
+            $products_with_sale  = Products::has('sales')
+                ->with('firstAttachments')
+                ->with('sales')
+                ->orderBy('created_at','desc')
+                ->get();
+
+            return $this->showCollection(WebProductResource::collection($products_with_sale));
         }
 
         return null;
     }
 
+
     /**
      * Store a newly created resource in storage.
      *
-     * @param Request $request
+     * @param StoreSale $request
+     * @param Products $employee_product
      * @return SaleResource|JsonResponse|Response
      */
-    public function store(StoreSale $request)
+    public function store(StoreSale $request,int $employee_product)
     {
-        $fieldName = $request->attributes();
+
+       // dd('sd');
+
+
         if(request()->expectsJson() && request()->acceptsJson()){
+            $product = Products::findOrFail($employee_product);
 
-            $product = Products::findOrFail($request->get($fieldName['product_id']));
+            if (!$product->sales()->exists()){
 
-            $startDate = Carbon::make($request->get($fieldName['start']));
-            $endDate = Carbon::make($request->get($fieldName['end']));
+                // frmat and validate dates
+                $startDate = Carbon::make($request->get('start'));
+                $endDate = Carbon::make($request->get('end'));
+                if($startDate > $endDate){
+                    return $this->errorResponse('End Sale Date must next the Start Date' ,422);
+                }
 
-            if($startDate > $endDate){
-                return $this->errorResponse('End Sale Date must next the Start Date' ,422);
+                // calc new price
+                $newPrice = 0;
+                $saleRate = 0;
+                if ($request->has('newPrice')){
+                    if ($request->get('newPrice') >= $product->price){
+                        return $this->errorResponse('New Price must be less than old' ,422);
+                    }
+
+                    $newPrice =  $request->get('newPrice');
+                    $saleRate = ($newPrice * 100) / $product->price;
+                }elseif ($request->has('saleRate')){
+
+                    $newPrice = ($product->price * (int) $request->get('saleRate')) / 100;
+                    $saleRate = $request->get('saleRate');
+                }
+
+                $newSale = Sales::create([
+                    'product_id' => $employee_product,
+                    'saleRate' => $saleRate,
+                    'newPrice' => $newPrice,
+                    'start' => $startDate,
+                    'end' => $endDate,
+                ]);
+
+
+            }else{
+                return $this->errorResponse('this product already have sale',422);
             }
-
-            $newPrice = ($product->price * (int) $request->get($fieldName['saleRate'])) / 100;
-            $newSale = Sales::create([
-                'product_id' => $request->get($fieldName['product_id']),
-                'saleRate' => $request->get($fieldName['saleRate']),
-                'newPrice' => $newPrice,
-                'start' => $startDate,
-                'end' => $endDate,
-
-            ]);
 
             return new SaleResource($newSale);
 
@@ -75,15 +115,24 @@ class SaleController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param Sales $sale
-     * @return SaleResource|Response
+     * @param int $product
+     * @param int $sale
+     * @return SaleResource|JsonResponse|Response
      * @throws Exception
      */
-    public function destroy(Sales $sale)
+    public function destroy(int $product, int $sale)
     {
         if (request()->expectsJson() && request()->acceptsJson()){
-            $sale->delete();
-            return new SaleResource($sale);
+           // dd($sale);
+            $sale = Sales::findOrFail($sale);
+            $sale_product_id = $sale->product_id;
+            if ($product === $sale_product_id){
+                $sale->delete();
+                return $this->successResponse(['message' => 'sale was deleted.'],200);
+            }else{
+                $this->errorResponse('this sale not for this product',422);
+            }
+
         }
         return null;
     }
